Handle threading & errors better around TopologyListeners/SSE
Previously, any TopologyListener that threw an exception would
cause all remaining listeners to not fire during the sequence.
If the error was persistent, all future firings would also abort
when the erroring Listener was fired, preventing correct usage
of all subsequent listeners.

Now we duplicate the listener list, execute them asynchronously
through an executor, and if one throws, we remove it from the
listener list.  (This may need to be revisited?)

In the SSE servlet, we also moved the keep-alive task to a
recurring scheduled executor in order to avoid thread proliferation.
Additionally, if an error occurs, the keep-alive task will
become unscheduled, and while it goes away, it also removes
the associated TopologyListener powering the SSE connection.