SWARM-279, SWARM-280, SWARM-281: Topology Improvements
Previously, registrations of services with the topology were
bound to the http socket-binding, completely not supporting
https.

Now, for the set of (http, https) that is enabled within the
application, one or both will be registered.

To support distinction between them, registrations now support
a set of tags, and each registration is tagged as either "http"
or "https".

Additionally, topology.js in the web-app has been modified to
allow for know the tags of each registration.

Also, Ribbon's IsSecure flag is respected, and if set to be
true, the ServerList is populated with service instances
that have the 'https' tag, instead of the 'http' tag.